platform2: Create span using unsigned length
A recent CL https://crrev.com/c/4219636 checks that span sizes are
unsigned.

This CL changes the type of the length of a span to be unsigned to meet
the requirements. It is required for removing backforward patch in
libchrome uprev r1103195.

BUG=None
TEST=CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4263549
Auto-Submit: Cici Ruan <cuicuiruan@google.com>
Commit-Queue: Cici Ruan <cuicuiruan@google.com>
Tested-by: Cici Ruan <cuicuiruan@google.com>

+++ crash-reporter/ec_collector.cc
@@ -88,8 +88,8 @@ bool ECCollector::Collect() {
              "\nERROR: ec_parse_panicinfo: Non-zero return value (%d).\n",
              result),
          "Raw data: ",
-         base::Base64Encode(
-             base::make_span(reinterpret_cast<uint8_t*>(data), len)),
+         base::Base64Encode(base::make_span(reinterpret_cast<uint8_t*>(data),
+                                            static_cast<size_t>(len))),
          "\n"});
   }
